Implement context-generic DelegateBuildUpdateClientMessage
