+++ crates/cosmos/cosmos-chain-components/src/components/cosmos_to_cosmos.rs
@@ -1,6 +1,7 @@
 use cgp_core::prelude::*;
 use hermes_relayer_components::chain::impls::delegate::queries::client_state::QueryAndDecodeClientStateVia;
 use hermes_relayer_components::chain::traits::message_builders::create_client::CreateClientMessageBuilderComponent;
+use hermes_relayer_components::chain::traits::message_builders::update_client::UpdateClientMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::queries::client_state::{
     AllClientStatesQuerierComponent, ClientStateQuerierComponent,
 };
@@ -8,6 +9,7 @@ use hermes_relayer_components::chain::traits::queries::consensus_state::Consensu
 use prost_types::Any;
 
 use crate::impls::client::create_client_message::BuildCosmosCreateClientMessage;
+use crate::impls::client::update_client_message::BuildCosmosUpdateClientMessage;
 use crate::impls::queries::consensus_state::QueryCosmosConsensusStateFromChainHandle;
 
 pub struct CosmosToCosmosComponents;
@@ -19,9 +21,11 @@ delegate_components! {
             AllClientStatesQuerierComponent,
         ]:
             QueryAndDecodeClientStateVia<Any>,
-        CreateClientMessageBuilderComponent:
-            BuildCosmosCreateClientMessage,
         ConsensusStateQuerierComponent:
             QueryCosmosConsensusStateFromChainHandle,
+        CreateClientMessageBuilderComponent:
+            BuildCosmosCreateClientMessage,
+        UpdateClientMessageBuilderComponent:
+            BuildCosmosUpdateClientMessage,
     }
 }
