+++ crates/cosmos/cosmos-relayer/src/chain/checks.rs
@@ -1,7 +1,9 @@
 use hermes_cli_components::any_client::contexts::any_counterparty::AnyCounterparty;
+use hermes_relayer_components::chain::traits::message_builders::update_client::CanBuildUpdateClientMessage;
 use hermes_relayer_components::chain::traits::queries::client_state::{
     CanQueryAllClientStates, CanQueryClientState,
 };
+use hermes_relayer_components::chain::traits::queries::consensus_state::CanQueryConsensusState;
 use hermes_relayer_components::transaction::traits::poll_tx_response::CanPollTxResponse;
 use hermes_relayer_components::transaction::traits::query_tx_response::CanQueryTxResponse;
 use hermes_relayer_components::transaction::traits::submit_tx::CanSubmitTx;
@@ -20,9 +22,11 @@ pub trait CanUseCosmosChain:
     + CanIbcTransferToken<CosmosChain>
     + CanBuildIbcTokenTransferMessage<CosmosChain>
     + CanQueryClientState<CosmosChain>
+    + CanQueryConsensusState<CosmosChain>
     + CanQueryAllClientStates<CosmosChain>
     + CanQueryClientState<AnyCounterparty>
     + CanQueryAllClientStates<AnyCounterparty>
+    + CanBuildUpdateClientMessage<CosmosChain>
     + CanSubmitTx
     + CanPollTxResponse
     + CanQueryTxResponse
