+++ crates/cosmos/cosmos-relayer/src/chain/impls/mod.rs
@@ -3,4 +3,3 @@ pub mod events;
 pub mod ext;
 pub mod fields;
 pub mod types;
-pub mod update_client_message;
