+++ crates/relayer/relayer-components/src/chain/impls/delegate/message_builders/mod.rs
@@ -1,1 +1,2 @@
 pub mod create_client;
+pub mod update_client;
