+++ crates/sovereign/sovereign-relayer/src/exts/cosmos/chain.rs
@@ -1,7 +1,6 @@
 use cgp_core::prelude::*;
 use hermes_cosmos_chain_components::components::delegate::DelegateCosmosChainComponents;
 use hermes_cosmos_relayer::chain::impls::connection_handshake_message::DelegateCosmosConnectionHandshakeBuilder;
-use hermes_cosmos_relayer::chain::impls::update_client_message::DelegateCosmosUpdateClientMessageBuilder;
 use hermes_cosmos_relayer::contexts::chain::CosmosChain;
 use hermes_relayer_components::chain::impls::delegate::queries::client_state::QueryAndDecodeClientStateVia;
 use hermes_sovereign_chain_components::cosmos::impls::sovereign_to_cosmos::client::create_client_message::BuildCreateSovereignClientMessageOnCosmos;
@@ -9,7 +8,7 @@ use hermes_sovereign_chain_components::cosmos::impls::sovereign_to_cosmos::clien
 use hermes_sovereign_chain_components::cosmos::impls::sovereign_to_cosmos::connection_handshake_message::BuildSovereignConnectionHandshakeMessageOnCosmos;
 use hermes_relayer_components::chain::traits::message_builders::connection_handshake::CanBuildConnectionHandshakeMessages;
 use hermes_relayer_components::chain::traits::message_builders::create_client::{CanBuildCreateClientMessage, CreateClientMessageBuilderComponent};
-use hermes_relayer_components::chain::traits::message_builders::update_client::CanBuildUpdateClientMessage;
+use hermes_relayer_components::chain::traits::message_builders::update_client::{CanBuildUpdateClientMessage, UpdateClientMessageBuilderComponent};
 use hermes_relayer_components::chain::traits::queries::client_state::{AllClientStatesBytesQuerierComponent, CanQueryClientState, ClientStateQuerierComponent};
 use hermes_wasm_client_components::types::client_state::WasmClientState;
 
@@ -30,6 +29,8 @@ delegate_components! {
             AllClientStatesBytesQuerierComponent,
         ]:
             QueryAndDecodeClientStateVia<WasmClientState>,
+        UpdateClientMessageBuilderComponent:
+            BuildUpdateSovereignClientMessageOnCosmos,
         CreateClientMessageBuilderComponent:
             BuildCreateSovereignClientMessageOnCosmos,
     }
@@ -44,12 +45,6 @@ pub trait CanUseSovereignMethodsOnCosmosChain:
 }
 
 impl CanUseSovereignMethodsOnCosmosChain for CosmosChain {}
-
-delegate_components! {
-    DelegateCosmosUpdateClientMessageBuilder {
-        SovereignChain: BuildUpdateSovereignClientMessageOnCosmos,
-    }
-}
 
 delegate_components! {
     DelegateCosmosConnectionHandshakeBuilder {
